Feat :sparkles:cli 현재폴더에서 이미지 S3저장추가
