+++ cli/controlers/login.js
@@ -2,6 +2,7 @@ const yargs = require("yargs");
 const inquirer = require("inquirer");
 const chalk = require("chalk");
 const axios = require("axios");
+const { directory } = require("./directory.js");
 
 module.exports = {
   // *로그인*
@@ -34,10 +35,11 @@ module.exports = {
             console.log(chalk.red("회원가입좀.."));
           } else {
             console.log(chalk.blue("✅ 로그인 성공"));
+            directory();
           }
-          console.log(data);
+          // console.log(data);
         });
-        console.log(answers);
+        // console.log(answers);
       });
   },
 };
