+++ cli/package.json
@@ -16,8 +16,11 @@
     "axios": "^0.26.1",
     "chalk": "^2.4.1",
     "commander": "^9.1.0",
+    "form-data": "^4.0.0",
     "inquirer": "^8.2.1",
     "jsonwebtoken": "^8.5.1",
+    "multer": "^1.4.4",
+    "multer-s3": "^2.10.0",
     "yargs": "^17.4.0"
   }
 }
