+++ server/config/s3.js
@@ -1,0 +1,24 @@
+const multer = require("multer");
+const multerS3 = require("multer-s3");
+const aws = require("aws-sdk");
+require("dotenv").config();
+
+const s3 = new aws.S3({
+  accessKeyId: process.env.S3_KEYID,
+  secretAccessKey: process.env.S3_PRIVATE_KEY,
+  region: process.env.REGION,
+});
+
+let upload = multer({
+  storage: multerS3({
+    s3: s3,
+    bucket: process.env.BUCKET_NAME,
+    contentType: multerS3.AUTO_CONTENT_TYPE,
+    acl: "public-read",
+    key: (req, file, cb) => {
+      cb(null, `userImg/${Date.now()}_${file.originalname}`);
+    },
+  }),
+});
+
+exports.upload = multer(upload);
