+++ server/index.js
@@ -6,6 +6,8 @@ const cors = require("cors");
 const cookieParser = require("cookie-parser");
 const controllers = require("./controllers");
 const app = express();
+const { upload } = require("./config/s3");
+
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(
@@ -64,6 +66,15 @@ app.get("/board/:boardId", controllers.mainboarddetail);
 app.post("/board/add", controllers.createboard);
 app.patch("/board/update/:userId/:boardId", controllers.updateboard);
 app.delete("/board/:userId/:boardId", controllers.deleteboard);
+
+// s3
+app.post("/upload", upload.single("userImg"), controllers.uploadS3);
+// upload.single("profile_picture"); // "파일명"
+// upload.array("profile_picture", 5), //이미지 최대 수를 입력
+//   upload.fields([
+//     { name: "profile_picture", maxCount: 1 },
+//     { name: "company_pictures", maxCount: 5 },
+//   ]);
 
 const HTTPS_PORT = process.env.HTTPS_PORT || 4000;
 // let server;
