+++ server/package.json
@@ -4,25 +4,27 @@
     "description": "",
     "main": "index.js",
     "scripts": {
-      "start": "nodemon index.js"
+        "start": "nodemon index.js"
     },
     "author": "",
     "license": "ISC",
     "devDependencies": {
-      "prettier": "2.5.1"
+        "prettier": "2.5.1"
     },
     "dependencies": {
-      "axios": "^0.26.0",
-      "cookie-parser": "^1.4.6",
-      "cors": "^2.8.5",
-      "dotenv": "^16.0.0",
-      "express": "^4.17.3",
-      "jsonwebtoken": "^8.5.1",
-      "mysql": "^2.18.1",
-      "mysql2": "^2.3.3",
-      "nodemon": "^2.0.15",
-      "sequelize": "^6.17.0",
-      "sequelize-cli": "^6.4.1"
+        "aws-sdk": "^2.1101.0",
+        "axios": "^0.26.0",
+        "cookie-parser": "^1.4.6",
+        "cors": "^2.8.5",
+        "dotenv": "^16.0.0",
+        "express": "^4.17.3",
+        "jsonwebtoken": "^8.5.1",
+        "multer": "^1.4.4",
+        "multer-s3": "^2.10.0",
+        "mysql": "^2.18.1",
+        "mysql2": "^2.3.3",
+        "nodemon": "^2.0.15",
+        "sequelize": "^6.17.0",
+        "sequelize-cli": "^6.4.1"
     }
-  }
-  
+}
